+++ scoring/src/fll/web/report/CategoryScoresByJudge.java
@@ -102,7 +102,7 @@ public class CategoryScoresByJudge extends BaseFLLServlet {
             final String judge = judgesRS.getString(1);
 
             writer.write("<h3>"
-                + categoryTitle + " Division: " + division + " Judge: " + judge + "</h3>");
+                + categoryTitle + " Award Group: " + division + " Judge: " + judge + "</h3>");
 
             writer.write("<table border='0'>");
             writer.write("<tr><th colspan='3'>Team # / Organization / Team Name</th><th>Raw Score</th><th>Scaled Score</th></tr>");