+++ scoring/src/fll/web/report/RankingReport.java
@@ -107,11 +107,11 @@ public class RankingReport extends BaseFLLServlet {
         para.add(new Chunk(team.getTeamName()
             + " / " + team.getOrganization(), TITLE_FONT));
         para.add(Chunk.NEWLINE);
-        para.add(new Chunk("Division: "
+        para.add(new Chunk("Award Group: "
             + team.getEventDivision(), TITLE_FONT));
         para.add(Chunk.NEWLINE);
         para.add(new Chunk(
-                           "Each team is ranked in each category in the judging group and division they were judged in. Performance scores are ranked by division only. Teams may have the same rank if they were tied.",
+                           "Each team is ranked in each category in the judging group and award group they were judged in. Performance scores are ranked by division only. Teams may have the same rank if they were tied.",
                            RANK_VALUE_FONT));
         para.add(Chunk.NEWLINE);
         para.add(Chunk.NEWLINE);